Don't enqueue ingress for some service changes (#365)
When a service changes we get a service change event and an endpoint change event. Previously we'd enqueue the associated ingresses for all service changes. Now we only enqueue it when the service changes in a way that won't be covered by the endpoint changes.

Specifically for Service.spec.selector changes, which users sometimes use to do blue/green deployments, we will no longer enqueue an ingress - instead the changes will be handled (correctly) by the endpoint handler.