Fix ignoring JWK secrets in VS route policies
Previously, If a route in a VS referenced a JWT policy, the IC would
always mistakenly assume that the JWK secret that the policy
referenced didn't exist. As a result, the VS would be configured
with a warning about a missing secret and NGINX would return 500
status codes for requests to that VS.

This commit fixes that bug.

The bug was introduced in 253c3e9