Don't require default server TLS secret
If the default server TLS secret is not configured via
-default-server-tls-secret cli arg or it is not present on the filesystem
at /etc/nginx/secrets/default, the Ingress Controller will configure
NGINX to reject TLS connections to the default server.

Note: The default server is used in NGINX configuration to handle HTTP
and HTTPS request to hosts that are not configured by any Ingress,
VirtualServer or TransportServer resources. The default server simply
returns 404 responses.