Reload only once during the start
During the start, the IC processes resources in the cluster one by one
Each processing might lead to a new generated config or an updated config,
which requires an NGINX reload. After all resources are processed,
the IC pod is considered Ready (its readiness probe succeeds).

For clusters with big number of resources the combined time
of NGINX reloads during the start can be significant (minutes).

This commit ensures that the IC only reloads NGINX once during
the start, which reduces the start up time for clusters
with big number of resources.