Configure IC root filesystem as read-only (#3548)
Allow configuring IC root filesystem as read-only

The Nginx Ingress Controller has various protections against attacks,
such as running the service as non-root to avoid changes to files.
An additional industry best practice is having root filesystems set
as read-only so that the attack surface is further reduced by limiting
changes to binaries and libraries. This commit ensures such a defense
in depth is enabled for the Nginx Ingress Controller.

To accomplish read-only rootfs, we will create emptyDir Volumes for
- `/etc/nginx`
- `/var/cache/nginx`
- `/var/lib/nginx`
- `/var/log/nginx`

We populate `/etc` mount with an Init Container, as there are generic
configuration files here. If the base image is updated via Helm chart,
then Init Container will pull files from the new base image.

The other three empty volumes are populated during runtime, as they
are not expected to have valuable files as part of base image:
- `/var/cache` mount will be used in case caching is enabled.
- `/var/lib` is used for some sockets and temporary files.
- `/var/log` is used for logging in the default configuration.

By default, the emptyDir will be persisted as long as the Pod is not
disposed of. Simple Pod crashes will retain the data.

There is no maximum size limit implemented currently, but may be added
at some point in time as a further protection. For improved performance,
cache could be moved to ram tmpfs as well (currently backed by disk).

Do note, the initial commit does not set read-only root filesystem
as default. Instead, this is an opt-in feature available on the chart.
Users of non-Helm YAML manifests need to manually uncomment the
aforementioned volumes and initContainers.

Co-authored-by: Shaun <s.odonovan@f5.com>