Update theme version, adjust weight of top level pages & sections (#3979)
This commit updates the documentation theme version: with the change,
the "weight" frontmatter parameter for a handful of pages has also been
adjusted.

Prior to this theme version, folders would not respect the weight
frontmatter for the purposes of ordering page layout. This means that no
matter what the value was, folders would always be at the top in the
hierarchy.

With this change, that is no longer the case - and the adjustments to
existing pages ensure that they adhere to the structure the
documentation had previously.