Update manifest installation guide, fix theme highlighting & spec table (#4066)
* Update manifest installation guide, fix theme highlighting & spec table

This commit includes two small changes: the first is removing a line
from the Hugo configuration file that was preventing the appropriate
styling to occur for code blocks. The second is to fix an incorrectly
closed table in the technical specifications page.

The largest change is reformatting of the Installation with Manifests
page. Multiple sections of the page now have horizontal lines to
separate them visually: many instances of notes were replaced with the
formal Hugo shortcode for notes. Indentation was also fixed in multiple
places, with subheadings being used to create explicit flow where
previously it was implicit.

Finally, there are multiple instances of nouns being fixed, as well as
certain portions of text being rephrased for clarity or conciseness.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Fixes duplicated word in manifest install guide

Signed-off-by: Jodie Putrino <j.putrino@f5.com>

---------

Signed-off-by: Jodie Putrino <j.putrino@f5.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Jodie Putrino <j.putrino@f5.com>