Update psuedo-variable shortcodes (#6632)
Update psuedo-variable shortcodes

This commit fixes an issue with three shortcodes in the documentation
website: they are used as psuedo-variables for the sake of version
management. Newline characters were added to them erroneously, 
which was causing them to break tables, and may have also affected 
the formatting and rendering of documentation elsewhere.