allow customization of service http and https port names through helm (#7318)
* allow customization of service http and https port names through helm chart

This change allows users to customize the names of the NIC service's http and https ports through a helm chart. It adds defaults which are the same as the previous hardcoded values in the helm chart, thereby ensuring backwards compatibility.

This change will enable users of NIC to more smoothly integrate with NGINXaaS's AKS functionality. Our load balancer for kubernetes discovers the upstream name and upstream type that it needs to update by parsing service port names, so we need to allow users of our service to easily customize port names.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>