Remove documentation website files (#7915)
This commit removes the documentation website files, which have been
migrated to the main documentation repository. It also includes some
changes to configuration files that are redundant, such as removing
the deployment action from the repository.

Some configuration changes have been left in place for context, and
there will likely be some future follow up work to re-integrate certain
automations as part of enhancements to the documentation repository.

---

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>